+++ example/lib/main.dart
@@ -1,3 +1,6 @@
+import 'dart:async';
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:volume_controller/volume_controller.dart';
 
@@ -6,29 +9,39 @@ void main() {
 }
 
 class MyApp extends StatefulWidget {
+  const MyApp({super.key});
+
   @override
-  _MyAppState createState() => _MyAppState();
+  State<MyApp> createState() => _MyAppState();
 }
 
 class _MyAppState extends State<MyApp> {
-  double _volumeListenerValue = 0;
-  double _getVolume = 0;
-  double _setVolumeValue = 0;
+  late final VolumeController _volumeController;
+  late final StreamSubscription<double> _subscription;
+
+  double _currentVolume = 0;
+  double _volumeValue = 0;
+  bool _isMuted = false;
 
   @override
   void initState() {
     super.initState();
-    // Listen to system volume change
-    VolumeController().listener((volume) {
-      setState(() => _volumeListenerValue = volume);
-    });
 
-    VolumeController().getVolume().then((volume) => _setVolumeValue = volume);
+    _volumeController = VolumeController.instance;
+
+    // Listen to system volume change
+    _subscription = _volumeController.addListener((volume) {
+      setState(() => _volumeValue = volume);
+    }, fetchInitialVolume: true);
+
+    _volumeController
+        .isMuted()
+        .then((isMuted) => setState(() => _isMuted = isMuted));
   }
 
   @override
   void dispose() {
-    VolumeController().removeListener();
+    _subscription.cancel();
     super.dispose();
   }
 
@@ -41,7 +54,7 @@ class _MyAppState extends State<MyApp> {
         ),
         body: Column(
           children: [
-            Text('Current volume: $_volumeListenerValue'),
+            Text('Current volume: $_volumeValue'),
             Row(
               children: [
                 Text('Set Volume:'),
@@ -49,13 +62,9 @@ class _MyAppState extends State<MyApp> {
                   child: Slider(
                     min: 0,
                     max: 1,
-                    onChanged: (double value) {
-                      setState(() {
-                        _setVolumeValue = value;
-                        VolumeController().setVolume(_setVolumeValue);
-                      });
-                    },
-                    value: _setVolumeValue,
+                    onChanged: (double value) async =>
+                        await _volumeController.setVolume(value),
+                    value: _volumeValue,
                   ),
                 ),
               ],
@@ -63,46 +72,70 @@ class _MyAppState extends State<MyApp> {
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                Text('Volume is: $_getVolume'),
+                Text('Volume is: $_currentVolume'),
                 TextButton(
                   onPressed: () async {
-                    _getVolume = await VolumeController().getVolume();
+                    _currentVolume = await _volumeController.getVolume();
                     setState(() {});
                   },
                   child: Text('Get Volume'),
                 ),
               ],
             ),
-            TextButton(
-              onPressed: () => VolumeController().muteVolume(),
-              child: Text('Mute Volume'),
-            ),
-            TextButton(
-              onPressed: () => VolumeController().maxVolume(),
-              child: Text('Max Volume'),
-            ),
+            if (Platform.isAndroid || Platform.isIOS)
+              Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                  Text('Show system UI:${_volumeController.showSystemUI}'),
+                  TextButton(
+                    onPressed: () => setState(
+                      () => _volumeController.showSystemUI =
+                          !_volumeController.showSystemUI,
+                    ),
+                    child: Text('Show/Hide UI'),
+                  )
+                ],
+              ),
             Row(
               mainAxisAlignment: MainAxisAlignment.center,
               children: [
-                Text('Show system UI:${VolumeController().showSystemUI}'),
+                Text('Is Muted:$_isMuted'),
                 TextButton(
-                  onPressed: () => setState(() => VolumeController()
-                      .showSystemUI = !VolumeController().showSystemUI),
-                  child: Text('Show/Hide UI'),
-                )
+                  onPressed: () async {
+                    await updateMuteStatus(true);
+                  },
+                  child: Text('Mute'),
+                ),
+                TextButton(
+                  onPressed: () async {
+                    await updateMuteStatus(false);
+                  },
+                  child: Text('Unmute'),
+                ),
               ],
             ),
-            VolumeSlider(
-              display: Display.HORIZONTAL,
-              sliderActiveColor: Theme.of(context).primaryColor,
-              muteIconColor: Theme.of(context).colorScheme.surface,
-              upVolumeIconColor: Theme.of(context).colorScheme.surface,
-              visibleWidget: true,
-              sliderInActiveColor: Theme.of(context).dividerColor,
-            )
+            TextButton(
+              onPressed: () async {
+                _isMuted = await _volumeController.isMuted();
+                setState(() {});
+              },
+              child: Text('Update Mute Status'),
+            ),
           ],
         ),
       ),
     );
+  }
+
+  Future<void> updateMuteStatus(bool isMute) async {
+    await _volumeController.setMute(isMute);
+    if (Platform.isIOS) {
+      // On iOS, the system does not update the mute status immediately
+      // You need to wait for the system to update the mute status
+      await Future.delayed(Duration(milliseconds: 50));
+    }
+    _isMuted = await _volumeController.isMuted();
+
+    setState(() {});
   }
 }