+++ lib/volume_controller.dart
@@ -1,3 +1,1 @@
-library;
 export 'src/volume_controller.dart';
-export 'src/volume_slider.dart';